PSO - moved target-related components into a separate module
